requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
typer==0.20.0
requests==2.32.5
